Temporarily Revert "Generic Tracker interface, support for TB logging (#367)"
This reverts commit 3c48577282444ed7f288d1275b5dde4e84e92d7b.
