+++ README.md
@@ -149,8 +149,7 @@ model:
     gradient_checkpointing: true
     scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
-    type: wandb
+  wandb:
     project: "levanter"
     tags: [ "openwebtext", "gpt2"]
