+++ config/backpack.yaml
@@ -10,7 +10,7 @@ model:
   num_senses: 16
   sense_intermediate_scale: 4
 trainer:
-  tracker:
+  wandb:
     project: "levanter"
     tags: [ "openwebtext", "backpack" ]
