+++ config/gpt2_medium.yaml
@@ -8,7 +8,7 @@ model:
   gradient_checkpointing: true
   scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
+  wandb:
     project: "levanter"
     tags: [ "openwebtext", "gpt2"]
