+++ config/gpt2_small_fast.yaml
@@ -8,12 +8,9 @@ model:
   gradient_checkpointing: true
   scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
-    - type: wandb
-      project: "levanter"
-      tags: [ "openwebtext", "gpt2", "itest"]
-    - type: tensorboard
-      logdir: gs://levanter-checkpoints/tblogs/
+  wandb:
+    project: "levanter"
+    tags: [ "openwebtext", "gpt2", "itest"]
 
   mp: p=f32,c=bfloat16
   model_axis_size: 1
