+++ config/gpt2_small_fast_wiki.yaml
@@ -9,7 +9,7 @@ model:
   gradient_checkpointing: true
   scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
+  wandb:
     project: "levanter"
     tags: [ "openwebtext", "gpt2", "itest"]
