+++ config/llama2_7b.yaml
@@ -11,8 +11,7 @@ model:
 # initialize_from_hf: "meta-llama/Llama-2-7b-hf"
 # use_hf_model_config: true
 trainer:
-  tracker:
-    type: wandb
+  wandb:
     project: "levanter"
     tags: ["openwebtext", "llama"]
