+++ config/llama2_7b_continued.yaml
@@ -6,8 +6,7 @@ model:
 initialize_from_hf: true
 use_hf_model_config: true
 trainer:
-  tracker:
-    type: wandb
+  wandb:
     project: "levanter"
     tags: ["pile", "llama2"]
