+++ config/mpt_7b_continued.yaml
@@ -4,7 +4,7 @@ model:
 initialize_from_hf: true
 use_hf_model_config: true
 trainer:
-  tracker:
+  wandb:
     project: "levanter"
     tags: ["pile", "mpt"]
