+++ config/mpt_7b_continued_biomedlm.yaml
@@ -10,7 +10,7 @@ model:
 initialize_from_hf: "mosaicml/mpt-7b@68e1a8e0ebb9b30f3c45c1ef6195980f29063ae2"
 use_hf_model_config: true
 trainer:
-  tracker:
+  wandb:
     project: "levanter"
     tags: ["pubmed", "mpt", "continued"]
