+++ docs/Configuration-Guide.md
@@ -35,8 +35,7 @@ model:
   gradient_checkpointing: true
   scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
-    type: wandb
+  wandb:
     project: "levanter"
     tags: [ "openwebtext", "gpt2"]
 
@@ -179,34 +178,12 @@ The default step-based checkpoint policy is to save a checkpoint every 10,000 st
 
 
 
-## Trackers and Logging
+## WandB
 
+We mostly use wandb for logging, including using wandb for allocating the run id. We may change this.
 
-We mostly use [W&B](https://wandb.ai/site) for tracking values and other metadata about a run. However, we also support
-Tensorboard and a few other trackers. You can also use multiple trackers at once, or even write your own.
-See  [Trackers](dev/Trackers.md) for more information.
-
-### W&B
-
-Wandb is the default tracker and is installed by default. To use it, you can configure it in your config file:
-
-```yaml
-trainer:
-    tracker:
-        type: wandb
-        project: my-project
-        entity: my-entity
-```
-
-Because wandb is the default, you can also just do:
-
-```yaml
-trainer:
-    tracker:
-      project: my-project
-      entity: my-entity
-```
-
+These all live in a nested object `wandb` inside `trainer`. Most of these are the same as the corresponding `wandb.init`
+parameters.
 
 
 | Parameter      | Description                                    | Default                    |
@@ -228,35 +205,6 @@ of your main script.
 To use it, you must also set the right environment variables. Something like `XLA_FLAGS="--xla_dump_to=/tmp/output_folder/xla_dumps --xla_dump_hlo_pass_re=.*`.
 We will automatically parse out the env variable.
 
-### Tensorboard
-
-Tensorboard is also supported. To use it, you can configure it in your config file:
-
-```yaml
-trainer:
-    tracker:
-        type: tensorboard
-        logdir: logs
-```
-
-### Multiple Trackers
-
-In some cases, you may want to use multiple trackers at once.
-For example, you may want to use both W&B and Tensorboard.
-
-To do this, you can use the [levanter.tracker.tracker.CompositeTracker][] class, or, if using a config file, you
-can specify multiple trackers:
-
-```yaml
-trainer:
-  tracker:
-    - type: wandb
-      project: my-project
-      entity: my-entity
-    - type: tensorboard
-      logdir: logs
-```
-
 ## Ray Config
 
 Levanter will by default automatically start a Ray cluster with all
@@ -264,11 +212,11 @@ the machines being used for training. This is useful for distributed
 preprocessing. You can disable this behavior using `auto_start_cluster: false`.
 
 
-| Parameter            | Description                                                             | Default |
-|----------------------|-------------------------------------------------------------------------|---------|
-| `address`            | The address of the Ray cluster to connect to.                           | `None`  |
-| `start_workers`      | Whether to start Ray workers. If `False`, you must start them yourself. | `True`  |
-| `auto_start_cluster` | Whether to start a Ray cluster automatically.                           | `True`  |
+| Parameter           | Description                                                                 | Default |
+|---------------------|-----------------------------------------------------------------------------|---------|
+| `address`           | The address of the Ray cluster to connect to.                                | `None`  |
+| `start_workers`     | Whether to start Ray workers. If `False`, you must start them yourself.      | `True`  |
+| `auto_start_cluster`| Whether to start a Ray cluster automatically.                                | `True`  |
 
 
 ## Distributed Config
@@ -278,12 +226,12 @@ If you're not using SLURM or TPUs, you can specify the cluster manually using th
 
 **Don't use this on TPU, and possibly not on SLURM either.**
 
-| Parameter             | Description                                                               | Default                 |
-|-----------------------|---------------------------------------------------------------------------|-------------------------|
-| `coordinator_address` | The address of the coordinator. If `None`, we'll use the default address. | `None`                  |
-| `num_processes`       | The number of processes in the cluster.                                   | `None`                  |
-| `process_id`          | The process id of this process.                                           | `None`                  |
-| `local_device_ids`    | The local device ids of this process.                                     | ${CUDA_VISIBLE_DEVICES} |
+| Parameter           | Description                                                                 | Default                 |
+|---------------------|-----------------------------------------------------------------------------|-------------------------|
+| `coordinator_address`| The address of the coordinator. If `None`, we'll use the default address.   | `None`                  |
+| `num_processes`     | The number of processes in the cluster.                                     | `None`                  |
+| `process_id`        | The process id of this process.                                             | `None`                  |
+| `local_device_ids`  | The local device ids of this process.                                       | ${CUDA_VISIBLE_DEVICES} |
 
 
 
@@ -328,26 +276,8 @@ We won't go into detail here. You can see the auto-generated docs below.
 
 ::: levanter.checkpoint.Checkpointer
 
-### Trackers and Metrics
-
-See also [Trackers](dev/Trackers.md) for more information. Basic configuration is shown below.
-
-#### Single Tracker
-
-```yaml
-trainer:
-  tracker:
-    type: wandb
-    project: my-project
-    entity: my-entity
-```
-
-
-
-::: levanter.tracker.wandb.WandbConfig
-
-::: levanter.tracker.tensorboard.TensorboardConfig
-
+### Wandb
+::: levanter.logging.WandbConfig
 
 ### Distributed and Ray
