+++ docs/Training-On-Your-Data.md
@@ -214,8 +214,7 @@ model:
   gradient_checkpointing: true
   scale_attn_by_inverse_layer_idx: true
 trainer:
-  tracker:
-    type: wandb
+  wandb:
     project: "levanter" # TODO
     tags: ["gpt2"]
