+++ examples/alpaca/alpaca.py
@@ -217,7 +217,9 @@ def train(config: TrainArgs):
     def compute_loss(model: LmHeadModel, example: LmExample, key=None):
         return model.compute_loss(example, key=key).scalar()
 
-    with Trainer(config.trainer, optimizer, compute_loss) as trainer:
+    trainer = Trainer(config.trainer, optimizer, compute_loss)
+
+    with trainer.device_mesh:
         # how we shard parameters across devices
         parameter_axis_mapping = trainer.parameter_axis_mapping
