+++ src/levanter/__init__.py
@@ -3,6 +3,4 @@ import levanter.config as config
 import levanter.data as data
 import levanter.distributed as distributed
 import levanter.logging as logging
-import levanter.tracker as tracker
 import levanter.visualization as visualization
-from levanter.tracker import current_tracker, get_tracker, log_metrics, log_summary
