+++ src/levanter/data/shard_cache.py
@@ -1,7 +1,7 @@
 # Dataset for preprocessing data, tokenizing, and caching to disk.
 import asyncio
 import dataclasses
-import logging as pylogging
+import logging
 import os
 import sys
 import threading
@@ -31,6 +31,7 @@ import pyarrow as pa
 import pyarrow.parquet as pq
 import ray
 import tblib
+import wandb
 from dataclasses_json import dataclass_json
 from fsspec import AbstractFileSystem
 from ray.actor import ActorHandle
@@ -45,9 +46,6 @@ from rich.progress import (
     TimeRemainingColumn,
 )
 
-import levanter.tracker
-
-from .. import logging
 from . import ShardableDataset
 from ._preprocessor import BatchProcessor, BatchResult, as_record_batch, dict_from_record_batch
 from .sharded_dataset import ShardedDataset
@@ -57,7 +55,7 @@ T = TypeVar("T")
 T_co = TypeVar("T_co", covariant=True)
 _ExcInfo = Tuple[Optional[BaseException], tblib.Traceback]
 
-logger = pylogging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 DEFAULT_ROWS_PER_CHUNK = 1024 * 32
 LEDGER_FILE_NAME = "cache_ledger.json"
@@ -267,7 +265,7 @@ def _produce_cache_for_shard(
     """Produces chunks of preprocessed data from a single shard and writes them to disk. Chunks are written to sink,
     which is an actor of ChunkCacheBuilder."""
     # TODO: thread logging level through calls
-    pylogging.basicConfig(level=pylogging.INFO)
+    logging.basicConfig(level=logging.INFO)
     # load or create shard metadata (for recovery)
     try:
         shard_name = source.shard_names[shard_idx]
@@ -477,7 +475,7 @@ class RichMetricsMonitor(MetricsMonitor):
         self.progress.start()
 
 
-class LoggingMetricsMonitor(MetricsMonitor):
+class WandbMetricsMonitor(MetricsMonitor):
     last_metrics: Optional[InProgressCacheMetrics]
     last_time: Optional[float]
 
@@ -519,16 +517,16 @@ class LoggingMetricsMonitor(MetricsMonitor):
         self.last_metrics = metrics
         self.last_time = time.time()
 
-        levanter.tracker.log_metrics(to_log, step=None, commit=self.commit)
+        wandb.log(to_log, commit=self.commit)
 
 
 class LoggerMetricsMonitor(MetricsMonitor):
     # TODO: I'd like to get the trainer pbar migrated to rich and just use rich everywhere, but until then,
     # we have separate logging
-    def __init__(self, logger: Optional[Union[pylogging.Logger, str]] = None, level=pylogging.INFO):
+    def __init__(self, logger: Optional[Union[logging.Logger, str]] = None, level=logging.INFO):
         if isinstance(logger, str):
-            logger = pylogging.getLogger(logger)
-        self.logger = logger or pylogging.getLogger(__name__)
+            logger = logging.getLogger(logger)
+        self.logger = logger or logging.getLogger(__name__)
         self.level = level
 
     def __call__(self, metrics: InProgressCacheMetrics):
@@ -572,7 +570,7 @@ class _ShardStatus:
 def _mk_process_task(processor: BatchProcessor[T]):
     @ray.remote(num_cpus=processor.num_cpus, num_gpus=processor.num_gpus, resources=processor.resources)
     def process_task(batch: List[T]) -> pa.RecordBatch:
-        pylogging.basicConfig(level=pylogging.INFO)
+        logging.basicConfig(level=logging.INFO)
         return processor(batch)
 
     return process_task
@@ -581,7 +579,7 @@ def _mk_process_task(processor: BatchProcessor[T]):
 def _mk_queue_aware_process_task(processor: BatchProcessor[T], queue: ActorHandle):
     @ray.remote(num_cpus=processor.num_cpus, num_gpus=processor.num_gpus, resources=processor.resources)
     def process_task(batch: List[T]) -> pa.RecordBatch:
-        pylogging.basicConfig(level=pylogging.INFO)
+        logging.basicConfig(level=logging.INFO)
         ray.get(queue.task_running.remote())
         result = processor(batch)
         del batch
@@ -676,7 +674,7 @@ class ChunkCacheBuilder:
         processor: BatchProcessor[T],
         rows_per_chunk: int,
     ):
-        pylogging.basicConfig(level=pylogging.INFO)
+        logging.basicConfig(level=logging.INFO)
         self.broker_ref = broker_ref
         self.shard_status: Dict[str, _ShardStatus] = dict()
         self._current_round_robin = []
@@ -815,7 +813,7 @@ class ChunkCacheBroker:
     _finished_promise: asyncio.Future[None]
 
     def __init__(self, cache_dir: str, source: ShardedDataset[T], processor: BatchProcessor[T], rows_per_chunk: int):
-        pylogging.basicConfig(level=pylogging.INFO)
+        logging.basicConfig(level=logging.INFO)
         self.chunks = []
         self._reader_promises = {}
         self._is_finished = False
