+++ src/levanter/main/cache_dataset.py
@@ -1,13 +1,14 @@
 import logging
 import os
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+
+import wandb
 
 import levanter
-from levanter.data.shard_cache import LoggingMetricsMonitor, RichMetricsMonitor, build_cache
+from levanter.data.shard_cache import RichMetricsMonitor, WandbMetricsMonitor, build_cache
 from levanter.data.text import BatchTokenizer, LMDatasetConfig
 from levanter.distributed import RayConfig
-from levanter.logging import init_logging
-from levanter.tracker import NoopConfig, TrackerConfig
+from levanter.logging import init_logger
 
 
 logger = logging.getLogger(__name__)
@@ -15,16 +16,18 @@ logger = logging.getLogger(__name__)
 
 @dataclass
 class RayCachedLMDatasetConfig(LMDatasetConfig, RayConfig):
-    tracker: TrackerConfig = field(default_factory=NoopConfig)
+    pass
 
 
 @levanter.config.main()
 def main(args: RayCachedLMDatasetConfig):
     """Caches two different kinds of datasets. It can cache a dataset from a list of urls, or a dataset from a hf dataset"""
-    init_logging("cache_dataset.log")
+    init_logger("cache_dataset.log")
     args.initialize()
 
     tokenizer = args.the_tokenizer
+
+    wandb.init(mode="offline")
 
     for split in ["train", "validation"]:
         print(f"Caching {split} to {args.cache_dir}.")
@@ -37,7 +40,7 @@ def main(args: RayCachedLMDatasetConfig):
             logger.warning(f"Skipping {split} because it is empty.")
             continue
 
-        monitors = [RichMetricsMonitor(source.num_shards), LoggingMetricsMonitor("preprocess/" + split, commit=True)]
+        monitors = [RichMetricsMonitor(source.num_shards), WandbMetricsMonitor("preprocess/" + split, commit=True)]
 
         cache = build_cache(
             cache_dir=split_cache_dir,
