+++ src/levanter/main/lora_lm.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass, field
 from typing import Optional
 
 import jax.random as jrandom
+import wandb
 
 import haliax.random
 
@@ -94,14 +95,8 @@ def main(config: LoraLmConfig):
         all_param_count = parameter_count(state.model)
         just_lora_params = parameter_count(trainer.trainable_params_only(state.model))
 
-        levanter.tracker.log_summary(
-            {
-                "parameter_count": all_param_count,
-                "trainable_parameter_count": just_lora_params,
-                "fraction_trainable": just_lora_params * 1.0 / all_param_count,
-            }
-        )
-
+        wandb.summary["parameter_count"] = all_param_count
+        wandb.summary["trainable_parameter_count"] = just_lora_params
         logger.info(f"Total parameter count: {all_param_count}")
         logger.info(f"Trainable parameter count: {just_lora_params}")
         logger.info(f"Fraction of parameters that are trainable: {just_lora_params * 1.0 / all_param_count%.3}")
