--- src/levanter/tracker/__init__.py
@@ -1,16 +1,0 @@
-from levanter.tracker.helpers import log_optimizer_hyperparams
-from levanter.tracker.tracker import CompositeTracker, NoopConfig, NoopTracker, Tracker, TrackerConfig
-from levanter.tracker.tracker_fns import current_tracker, get_tracker, jit_log_metrics, log_metrics, log_summary
-
-
-__all__ = [
-    "Tracker",
-    "TrackerConfig",
-    "CompositeTracker",
-    "log_optimizer_hyperparams",
-    "NoopTracker",
-    "current_tracker",
-    "jit_log_metrics",
-    "log_metrics",
-    "log_summary",
-]
