--- src/levanter/tracker/tensorboard.py
@@ -1,86 +1,0 @@
-import logging
-import os
-import typing
-from dataclasses import dataclass
-from typing import Any, Optional
-
-from levanter.tracker import Tracker, TrackerConfig, helpers
-
-
-pylogger = logging.getLogger(__name__)
-
-if typing.TYPE_CHECKING:
-    from tensorboardX import SummaryWriter  # noqa: F401
-
-
-class TensorboardTracker(Tracker):
-    name: str = "tensorboard"
-
-    def __init__(self, writer: "SummaryWriter"):
-        self.writer = writer
-
-    def log_hyperparameters(self, hparams: dict[str, Any]):
-        self.writer.add_hparams(hparams, {"dummy": 0})
-
-    def log(self, metrics: dict[str, Any], *, step, commit=None):
-        del commit
-        for k, v in metrics.items():
-            self.writer.add_scalar(k, v, step)
-
-    def log_summary(self, metrics: dict[str, Any]):
-        for k, v in metrics.items():
-            self.writer.add_scalar(k, v, global_step=None)
-
-    def log_artifact(self, artifact, *, name: Optional[str] = None, type: Optional[str] = None):
-        pylogger.warning("TensorboardLogger does not support logging artifacts yet")
-        pass
-
-
-@TrackerConfig.register_subclass("tensorboard")
-@dataclass
-class TensorboardConfig(TrackerConfig):
-    logdir: str = "tblogs"
-    comment: Optional[str] = ""
-    purge_step: Optional[int] = None
-    max_queue: Optional[int] = 10
-    flush_secs: Optional[int] = 120
-    filename_suffix: Optional[str] = ""
-    write_to_disk: Optional[bool] = True
-
-    def init(self, run_id: Optional[str], hparams=None) -> TensorboardTracker:
-        dir_to_write = self.logdir
-        if run_id is not None:
-            dir_to_write = os.path.join(dir_to_write, run_id)
-
-        pylogger.info(f"Writing Tensorboard logs to {dir_to_write}")
-
-        from tensorboardX import SummaryWriter  # noqa: F811
-
-        writer = SummaryWriter(
-            dir_to_write,
-            comment=self.comment,
-            purge_step=self.purge_step,
-            max_queue=self.max_queue,
-            flush_secs=self.flush_secs,
-            filename_suffix=self.filename_suffix,
-            write_to_disk=self.write_to_disk,
-        )
-
-        hparams_dict = helpers.hparams_to_dict(hparams)
-        hparams_dict = _flatten_nested_dict(hparams_dict)
-
-        writer.add_hparams(hparams_dict, {"dummy": 0})
-
-        return TensorboardTracker(writer)
-
-
-def _flatten_nested_dict(d):
-    def items():
-        for key, value in d.items():
-            if isinstance(value, dict):
-                for subkey, subvalue in _flatten_nested_dict(value).items():
-                    yield key + "/" + subkey, subvalue
-            else:
-                yield key, value
-
-    return dict(items())
