+++ tests/test_logging.py
@@ -3,7 +3,7 @@ import pathlib
 import pytest
 from git import InvalidGitRepositoryError, NoSuchPathError, Repo
 
-from levanter.tracker.helpers import infer_experiment_git_root
+from levanter.logging import WandbConfig
 
 
 def test_infer_experiment_git_root():
@@ -13,7 +13,7 @@ def test_infer_experiment_git_root():
     except (InvalidGitRepositoryError, NoSuchPathError):
         pytest.skip("test not running in a git repo")
 
-    root = infer_experiment_git_root()
+    root = WandbConfig._infer_experiment_git_root()
 
     # ensure that 1) this is a git root and 2) this source file is underneath
     assert root is not None
