--- tests/test_tracker.py
@@ -1,80 +1,0 @@
-# NOTE: Do not explicitly import wandb/other trackers here, as this will cause the tests to trivially pass.
-import dataclasses
-from typing import Tuple
-
-import pytest
-import yaml
-
-import levanter.tracker
-from levanter.tracker import CompositeTracker, TrackerConfig
-
-
-def test_tracker_plugin_stuff_works():
-    assert TrackerConfig.get_choice_class("wandb") is not None
-    with pytest.raises(KeyError):
-        TrackerConfig.get_choice_class("foo")
-
-
-def test_tracker_plugin_default_works():
-    config = """
-    tracker:
-        entity: foo
-    """
-    parsed = yaml.safe_load(config)
-
-    @dataclasses.dataclass
-    class ConfigHolder:
-        tracker: TrackerConfig
-
-    import draccus
-
-    tconfig = draccus.decode(ConfigHolder, parsed).tracker
-
-    assert isinstance(tconfig, TrackerConfig.get_choice_class("wandb"))
-
-    assert tconfig.entity == "foo"  # type: ignore
-
-
-def test_tracker_plugin_multi_parsing_work():
-    config = """
-    tracker:
-        type: noop
-    """
-    parsed = yaml.safe_load(config)
-
-    @dataclasses.dataclass
-    class ConfigHolder:
-        tracker: TrackerConfig | Tuple[TrackerConfig, ...]
-
-    import draccus
-
-    from levanter.tracker.tracker import NoopConfig
-
-    assert isinstance(draccus.decode(ConfigHolder, parsed).tracker, NoopConfig)
-
-    config = """
-    tracker:
-        - type: noop
-        - type: wandb
-    """
-    parsed = yaml.safe_load(config)
-    decoded = draccus.decode(ConfigHolder, parsed).tracker
-    assert decoded == (NoopConfig(), TrackerConfig.get_choice_class("wandb")())
-
-
-def test_get_tracker_by_name():
-    wandb_config = TrackerConfig.get_choice_class("wandb")
-    if wandb_config is None:
-        pytest.skip("wandb not installed")
-
-    from levanter.tracker import NoopTracker
-
-    wandb1 = wandb_config(mode="disabled").init(None)
-    tracker = CompositeTracker([wandb1, NoopTracker()])
-
-    with tracker:
-        assert levanter.tracker.get_tracker("wandb") is wandb1
-        assert levanter.tracker.get_tracker("noop") is not None
-
-        with pytest.raises(KeyError):
-            levanter.tracker.get_tracker("foo")
