+++ tests/test_train_lm.py
@@ -8,7 +8,7 @@ import ray
 import levanter.main.train_lm as train_lm
 import tiny_test_corpus
 from levanter.distributed import RayConfig
-from levanter.tracker.wandb import WandbConfig
+from levanter.logging import WandbConfig
 from levanter.utils.py_utils import logical_cpu_core_count
